refactor(web): Extract asset grid layout component from TimelineDateGroup and split into AssetLayout and Month components
